Move LayoutAllocators to the layouts package
