+++ core/shared/src/main/scala/eu/joaocosta/interim/skins/ButtonSkin.scala
@@ -2,7 +2,6 @@ package eu.joaocosta.interim.skins
 
 import eu.joaocosta.interim.TextLayout._
 import eu.joaocosta.interim._
-import eu.joaocosta.interim.api.LayoutAllocator
 import eu.joaocosta.interim.api.Primitives._
 
 trait ButtonSkin:
